requirements.txt
asgiref==3.11.0
certifi==2026.1.4
charset-normalizer==3.4.4
Django==6.0.1
django-cors-headers==4.9.0
django-mailjet==0.3.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
mailjet-rest==1.5.1
PyJWT==2.10.1
requests==2.32.5
sqlparse==0.5.5
urllib3==2.6.3